Restructure web components
