+++ web/src/components/glossary.tsx
